Split TaskInstance details component
